Plans by Opus

+++ Engine/Source/Frame/Collections/AreaLights/AreaLightsRender.cpp
@@ -80,7 +80,7 @@ void AreaLightsInterpolate::Render([[maybe_unused]] const game::FrameInterpolate
 		// Calculate center and get type configuration
 		XMVECTOR vecCenter = XMVectorScale(XMVectorAdd(XMVectorAdd(vecVisiblePos0, vecVisiblePos1), XMVectorAdd(vecVisiblePos2, vecVisiblePos3)), 0.25f);
 		const AreaLightsType& rType = AreaLightsInterpolate::GetType(rCurrent.puiTypeIndices[i]);
-		float fTextureIndex = gpTextureManager->mTextureDescriptors.CrcToIndex(rType.crc);
+		float fTextureIndex = static_cast<float>(gpTextureManager->mTextureDescriptors.CrcToIndex(rType.crc));
 		float fBlurredTextureIndex = gpTextureManager->mTextureDescriptors.CrcToBlurredIndex(rType.crc);
 		float fIntensityMultiplier = rCurrent.pfIntensityMultipliers[i];
 		float fVisibleIntensity = rType.pVisibleIntensityWrapper != nullptr ? rType.pVisibleIntensityWrapper->Get() : rType.fVisibleIntensity;
